requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-openai
langchain-groq
openai
requests
python-dotenv
plotly
langsmith
certifi